Initial commit of assg 13 on C++ Classes
Initial commit of assignment 13 template
project.  This assignment has students declare and
implement a C++ class `Deck` to represent
a deck of standard playing cards, and implement
some constructors and member methods of the class.

This implementation should be complete barring
typos or bugs.  3 tasks defined, and stubs/code
and instructions in place for assignment.